Refactor our font stack (#99429)
* fonts - adopt monaco-monospace-font in more places and make it consistent

* font - use platform specific fonts in more places

* font - move system-ui before Ubuntu font in places where we cannot control platform

* font - only use system-ui on Linux

* fonts - adjust font stack for other windows
