+++ src/vs/code/browser/workbench/callback.html
@@ -35,7 +35,7 @@
 				display: flex;
 				flex-direction: column;
 				color: white;
-				font-family: "Segoe UI", "Helvetica Neue", "Helvetica", Arial, sans-serif;
+				font-family: -apple-system, BlinkMacSystemFont, "Segoe WPC", "Segoe UI", system-ui, "Ubuntu", "Droid Sans", sans-serif;
 				background-color: #373277;
 			}
